refactoring7
